Move callbacks to model from ChunkCell
